pyproject.toml
[project]
name = "contract-qa-high-precision-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=0.6.3",
    "cohere>=5.13.11",
    "coloredlogs>=15.0.1",
    "fastapi>=0.115.6",
    "ipywidgets>=8.1.5",
    "langchain>=0.3.15",
    "langchain-cohere>=0.3.5",
    "langchain-community>=0.3.15",
    "langchain-core>=0.3.31",
    "langchain-experimental>=0.3.4",
    "langchain-google-genai>=2.0.9",
    "langchain-openai>=0.3.1",
    "litellm>=1.59.1",
    "markdownify>=0.14.1",
    "nltk>=3.9.1",
    "openai>=1.59.9",
    "protobuf>=5.29.3",
    "pyautogen>=0.3.2",
    "python-docx>=1.1.2",
    "python-dotenv>=1.0.1",
    "python-levenshtein>=0.26.1",
    "semantic-router[processing]>=0.0.72",
    "sentence-transformers>=3.3.1",
    "streamlit>=1.41.1",
    "unstructured[all-docs]>=0.14.8",
    "uvicorn[standard]>=0.34.0",
    "weaviate-client>=4.10.4",
]
